HDF5 General Library Functions {rhdf5}	R Documentation

HDF5 General Library Functions

Description

These low level functions provide general general libary functions for HDF5.

Usage

H5open             ()
H5close            ()
H5garbage_collect  ()
H5get_libversion   ()

Details

These low level functions provide general general libary functions for HDF5 C-library libhdf5. See http://www.hdfgroup.org/HDF5/doc/RM/RM_H5.html for further details.

Value

H5open initializes the HDF5 library. H5close flushes all data to disk, closes all open identifiers, and cleans up memory. H5garbage_collect cleans up memory. H5get_libversion returns the version number of the HDF5 C-library.

Author(s)

Bernd Fischer

References

http://www.hdfgroup.org/HDF5

See Also

h5version, rhdf5

Examples

H5open()
H5close()
H5garbage_collect()
H5get_libversion()

[Package rhdf5 version 2.26.0 Index]