Create an attribute for an HDF5 object

Creates an attribute, name, which is attached to the object specified by the identifier h5obj. The attribute name must be unique for the object.

H5Acreate(h5obj, name, dtype_id, h5space)

Arguments

h5obj
An object of class H5IdComponent representing a H5 object identifier (file, group, or dataset). See H5Fcreate(), H5Fopen(), H5Gcreate(), H5Gopen(), H5Dcreate(), or H5Dopen() to create an object of this kind.

name
The name of the attribute (character).

dtype_id
A character name of a datatype. See h5const("H5T") for possible datatypes. Can also be an integer representing an HDF5 datatype. Only simple datatypes are allowed for atttributes.

h5space
An object of class H5IdComponent representing a H5 dataspace. See H5Dget_space(), H5Screate_simple(), H5Screate() to create an object of this kind.

Value

An object of class H5IdComponent representing a H5 attribute identifier.