Check whether an specific attribute exists for an HDF5 object

Check whether an specific attribute exists for an HDF5 object

Usage

H5Aexists(h5obj, name)

Arguments

h5obj

An object of class H5IdComponent representing a H5 object identifier (file, group, or dataset). See H5Fcreate(), H5Fopen(), H5Gcreate(), H5Gopen(), H5Dcreate(), or H5Dopen() to create an object of this kind.

name

The name of the attribute (character).

Value

A logical value indicating whether an attribute with name name exists for the object specified by h5obj.