Change the dimensions of an HDF5 dataset

Change the dimensions of an HDF5 dataset

H5Dset_extent(h5dataset, size)

Details

This function can only be applied to datasets that meet the following criteria:

A chunked dataset with unlimited dimensions

A chunked dataset with fixed dimensions if the new dimension sizes are less than the maximum sizes set with maxdims #'

Fields

h5dataset

Object of class H5IdComponent representing an open HDF5 dataset.

size

An integer vector with the new dimension of the dataset.

Author

Bernd Fischer, Mike Smith