HDF5 Group Interface {rhdf5}	R Documentation

HDF5 Group Interface

Description

The Group interface functions create and manipulate groups of objects in an HDF5 file.

Usage

H5Gcreate      (h5loc, name)
H5Gcreate_anon (h5loc)
H5Gopen        (h5loc, name)
H5Gclose       (h5group)
H5Gget_info    (h5loc)
H5Gget_info_by_idx  (h5loc, n, group_name = ".", 
                       index_type = h5default("H5_INDEX"), 
		       order = h5default("H5_ITER"))
H5Gget_info_by_name (h5loc, group_name)

Arguments

h5loc
An object of class H5IdComponent representing a H5 location identifier (file or group). See H5Fcreate, H5Fopen, H5Gcreate, H5Gopen to create an object of this kind.

name
Name of the group.

h5group
An object of class H5IdComponent representing a H5 group identifier. See H5Gcreate, H5Gopen to create an object of this kind.

n
Position in the index of the group for which information is retrieved (Counting is 1-based).

group_name
An additional group name specifying the group for which information is sought. It is interpreted relative to h5loc.

index_type
See h5const("H5_INDEX") for possible arguments.

order
See h5const("H5_ITER") for possible arguments.

Details

Interface to the HDF5 C-library libhdf5. See http://www.hdfgroup.org/HDF5/doc/RM/RM_H5G.html for further details.

Value

H5Gcreate, H5Gcreate_anon, and H5Gopen return an object of class H5IdComponent representing a H5 group identifier.

H5Gget_info, H5Gget_info_by_idx, and H5Gget_info_by_name return a list with the group information.

The other functions return the standard return value from their respective C-functions.

Author(s)

Bernd Fischer

References

http://www.hdfgroup.org/HDF5

See Also

rhdf5

Examples

fid <- H5Fcreate("ex_H5G.h5")
gid <- H5Gcreate(fid, "foo")
gid
H5Gget_info(gid)
H5Gclose(gid)

H5Gget_info_by_idx(fid,1)
H5Gget_info_by_name(fid,"foo")

H5Fclose(fid)

[Package rhdf5 version 2.26.0 Index]