Open a specified group

Open a specified group

Usage

H5Gopen(h5loc, name)

Arguments

h5loc

An object of class H5IdComponent representing a H5 file or group that contains the group to be opened.

name

Name of the group to open.

Value

An object of class H5IdComponent representing the opened group. When access to the group is no longer needed this should be released with H5Gclose() to prevent resource leakage.

See also