HDF5 Property List Interface {rhdf5}	R Documentation

HDF5 Property List Interface

Description

The functions, macros, and subroutines listed here are used to manipulate property list objects in various ways, including to reset property values. With the use of property lists, HDF5 functions have been implemented and can be used in applications with many fewer parameters than would be required without property lists.

Usage

H5Pcreate            (type = h5default("H5P"), native = FALSE)
H5Pcopy              (h5plist)
H5Pget_class         (h5plist)
H5Pclose             (h5plist)
H5Pclose_class       (h5plistclass)
H5Pequal             (h5plistclass1, h5plistclass2)
H5Pset_libver_bounds (h5plist, 
                      libver_low = "H5F_LIBVER_EARLIEST", 
                      libver_high = "H5F_LIBVER_LATEST")
H5Pget_libver_bounds (h5plist)

Arguments

type
A character name of a property list type. See h5const("H5P") for possible property list types. Can also be an integer representing an HDF5 property list type.

native
An object of class logical. If TRUE, array-like objects are treated as stored in HDF5 row-major rather than R column-major orientation. Using native = TRUE increases HDF5 file portability between programming languages. A file written with native = TRUE should also be read with native = TRUE

.

h5plist
An object of class H5IdComponent representing a H5 property list identifier. See H5Pcreate or H5Pcopy to create an object of this kind.

h5plistclass, h5plistclass1, h5plistclass2
An object of class H5IdComponent representing a H5 property list class identifier. See H5Pget_class to create an object of this kind.

libver_low
A character value representing the lower bound on the library version for reading or writing the HDF5 file. See h5const("H5F_LIBVER") for possible arguments. Default is "H5F_LIBVER_EARLIEST"

libver_high
A character value representing the higher bound on the library version for reading or writing the HDF5 file. See h5const("H5F_LIBVER") for possible arguments. Default is "H5F_LIBVER_LATEST"

Details

Interface to the HDF5 C-library libhdf5. See http://www.hdfgroup.org/HDF5/doc/RM/RM_H5P.html for further details.

Value

H5Pcreate and H5Pcopy return an object of class H5IdComponent representing a H5 property list identifier.

H5Pget_class returns an object of class H5IdComponent representing a H5 property list class identifier.

The other functions return the standard return value from their respective C-functions.

Author(s)

Bernd Fischer

References

http://www.hdfgroup.org/HDF5

See Also

rhdf5, H5P_DATASET_CREATE

Examples

pid <- H5Pcreate()
pid2 <- H5Pcopy(pid)
pclid <- H5Pget_class(pid)
H5Pclose_class(pclid)
H5Pclose(pid)
H5Pclose(pid2)

[Package rhdf5 version 2.26.0 Index]