HDF5 Dataset Access Property List Interface {rhdf5}	R Documentation

HDF5 Dataset Access Property List Interface

Description

The functions, macros, and subroutines listed here are used to manipulate dataset access property list objects in various ways, including to reset property values. With the use of property lists, HDF5 functions have been implemented and can be used in applications with many fewer parameters than would be required without property lists.

Usage

H5Pset_chunk_cache                 ( h5plist, rdcc_nslots, rdcc_nbytes, rdcc_w0 )

Arguments

h5plist
An object of class H5IdComponent representing a H5 property list identifier of class H5P_DATASET_ACCESS. See H5Pcreate or H5Pcopy to create an object of this kind.

rdcc_nslots
An integer. The number of chunk slots in the raw data chunk cache for this dataset.

rdcc_nbytes
An integer. The total size of the raw data chunk cache for this dataset.

rdcc_w0
double. The chunk preemption policy for this dataset.

Details

Interface to the HDF5 C-library libhdf5. See http://www.hdfgroup.org/HDF5/doc/RM/RM_H5P.html for further details. See H5P for documentation of more functions to manupulate property lists of other property list classes.

Value

The functions return the standard return value from their respective C-functions.

Author(s)

Bernd Fischer

References

http://www.hdfgroup.org/HDF5

See Also

rhdf5, H5P

Examples

pid <- H5Pcreate("H5P_DATASET_ACCESS")
H5Pset_chunk_cache( pid, 100, 10000, 0.5)
H5Pclose(pid)

[Package rhdf5 version 2.26.0 Index]