HDF5 Dataset Create Property List Interface {rhdf5}	R Documentation

HDF5 Dataset Create Property List Interface

Description

The functions, macros, and subroutines listed here are used to manipulate dataset creation property list objects in various ways, including to reset property values. With the use of property lists, HDF5 functions have been implemented and can be used in applications with many fewer parameters than would be required without property lists.

Usage

H5Pset_layout                      ( h5plist, layout = h5default("H5D") )
H5Pget_layout                      ( h5plist )
H5Pset_chunk                       ( h5plist, dim )
H5Pget_chunk                       ( h5plist )
H5Pset_deflate                     ( h5plist, level )
H5Pset_fill_value                  ( h5plist, value )

H5Pfill_value_defined              ( h5plist )
H5Pset_fill_time                   ( h5plist, fill_time = h5default("H5D_FILL_TIME") )
H5Pget_fill_time                   ( h5plist )
H5Pset_alloc_time                  ( h5plist, alloc_time = h5default("H5D_ALLOC_TIME") )
H5Pget_alloc_time                  ( h5plist )

Arguments

h5plist
An object of class H5IdComponent representing a H5 property list identifier of class H5P_DATASET_CREATE. See H5Pcreate or H5Pcopy to create an object of this kind.

layout
A character name of a dataset layout type. See h5const("H5D") for possible property list types. Can also be an integer representing an HDF5 dataset layout type.

dim
The chunk size used to store the dataset. It is an integer vector of the same length as the dataset dims.

level
The compression level used. An integer value between 0 (no compression) and 9 (highest and slowest compression).

value
Standard value for filling the dataset. The storage.mode of value has to be convertable to the dataset type by HDF5.

fill_time
A character name of a H5D_FILL_TIME type. See h5const("H5D_FILL_TIME") for possible values. Can also be an integer representing an HDF5 H5D_FILL_TIME type.

alloc_time
A character name of a H5D_ALLOC_TIME type. See h5const("H5D_ALLOC_TIME") for possible property list types. Can also be an integer representing an HDF5 H5D_ALLOC_TIME type.

Details

Interface to the HDF5 C-library libhdf5. See http://www.hdfgroup.org/HDF5/doc/RM/RM_H5P.html for further details. See H5P for documentation of more functions to manupulate property lists of other property list classes.

Value

The functions return the standard return value from their respective C-functions.

Author(s)

Bernd Fischer

References

http://www.hdfgroup.org/HDF5

See Also

rhdf5, H5P

Examples

pid <- H5Pcreate("H5P_DATASET_CREATE")
H5Pset_fill_time( pid, "H5D_FILL_TIME_ALLOC" )
H5Pset_chunk(pid, c(1000,1,1))
H5Pset_deflate(pid, 6)

H5Pget_layout( pid )
H5Pfill_value_defined( pid )
H5Pget_fill_time( pid )
H5Pget_alloc_time( pid )
H5Pget_chunk( pid)

H5Pclose(pid)

[Package rhdf5 version 2.26.0 Index]