HDF5 Link Create Property List Interface {rhdf5}	R Documentation

HDF5 Link Create Property List Interface

Description

The functions, macros, and subroutines listed here are used to manipulate link creation property list objects in various ways, including to reset property values. With the use of property lists, HDF5 functions have been implemented and can be used in applications with many fewer parameters than would be required without property lists.

Usage

H5Pset_char_encoding               ( h5plist, encoding  = h5default("H5T_CSET"))
H5Pget_char_encoding               ( h5plist )
H5Pset_create_intermediate_group   ( h5plist, crt_intermed_group )
H5Pget_create_intermediate_group   ( h5plist )

Arguments

h5plist
An object of class H5IdComponent representing a H5 property list identifier of class H5P_LINK_CREATE. See H5Pcreate or H5Pcopy to create an object of this kind.

encoding
A character name of an encoding type. See h5const("H5T_CSET") for possible property list types. Can also be an integer representing an HDF5 encoding type.

crt_intermed_group
Logical. If TRUE intermediate groups are created.

Details

Interface to the HDF5 C-library libhdf5. See http://www.hdfgroup.org/HDF5/doc/RM/RM_H5P.html for further details. See H5P for documentation of more functions to manupulate property lists of other property list classes.

Value

The functions return the standard return value from their respective C-functions.

Author(s)

Bernd Fischer

References

http://www.hdfgroup.org/HDF5

See Also

rhdf5, H5P

Examples

pid <- H5Pcreate("H5P_LINK_CREATE")
H5Pset_char_encoding( pid, "H5T_CSET_ASCII")
H5Pget_char_encoding( pid)
H5Pset_create_intermediate_group(pid, TRUE)
H5Pget_create_intermediate_group(pid)
H5Pclose(pid)

[Package rhdf5 version 2.26.0 Index]