h5createGroup {rhdf5}	R Documentation

Create HDF5 group

Description

Creates a group within an HDF5 file.

Usage

h5createGroup (file, group)

Arguments

file
The filename (character) of the file in which the dataset will be located. For advanced programmers it is possible to provide an object of class H5IdComponent representing a H5 location identifier (file or group). See H5Fcreate, H5Fopen, H5Gcreate, H5Gopen to create an object of this kind.

group
The name of the new group. The name can contain a hierarchy of groupnames, e.g. 'group1/group2/newgroup', but the function will fail if the top level group do not exists.

Details

Creates a new group within an HDF5 file.

Value

Returns TRUE is group was created successfully and FALSE otherwise.

Author(s)

Bernd Fischer

References

http://www.hdfgroup.org/HDF5

See Also

h5createFile, h5createDataset, h5read, h5write, rhdf5

Examples

h5createFile("ex_createGroup.h5")

# create groups
h5createGroup("ex_createGroup.h5","foo")
h5createGroup("ex_createGroup.h5","foo/foobaa")

h5ls("ex_createGroup.h5")

[Package rhdf5 version 2.26.0 Index]