h5delete {rhdf5}	R Documentation

Delete contents of HDF5 file

Description

Deletes the specfied group or dataset from within an HDF5 file.

Usage

    h5delete( file, name )

Arguments

file
The filename (character) of the file in which the object is located.

name
The names of the object to be deleted.

Author(s)

Mike Smith

[Package rhdf5 version 2.26.0 Index]