h5listIdentifier {rhdf5}	R Documentation

list all valid H5 identifier.

Description

A list of all valid H5 identifier. H5 objects should be closed after usage to release resources.

Usage

h5listIdentifier()
h5validObjects(native = FALSE)

Arguments

native
An object of class logical. If TRUE, array-like objects are treated as stored in HDF5 row-major rather than R column-major orientation. Using native = TRUE increases HDF5 file portability between programming languages. A file written with native = TRUE should also be read with native = TRUE

.

Value

h5validObjects returns a list of H5IdComponent objects. h5listIdentifier prints the valid identifiers on screen and returns NULL.

Author(s)

Bernd Fischer

References

http://www.hdfgroup.org/HDF5

See Also

rhdf5

Examples

h5createFile("ex_list_identifier.h5")

# create groups
h5createGroup("ex_list_identifier.h5","foo")

h5listIdentifier()
h5validObjects()

[Package rhdf5 version 2.26.0 Index]