&-methods {rhdf5}	R Documentation

Methods for Function & in Package rhdf5

Description

Methods for function & in package rhdf5.

Methods

signature(e1 = "H5IdComponent", e2 = "ANY")

Returns a group handle or dataset handle for the group or dataset e2 in the HDF5 location e1. e1 can either be a file handle as returned by H5Fopen() or a group handle as e.g. returned by h5f$g1 or h5f$'/g1/g2's.

[Package rhdf5 version 2.26.0 Index]