[<–methods {rhdf5}	R Documentation

Methods for Function [<- in Package rhdf5

Description

Methods for function [<- in package rhdf5.

Methods

signature(x = "H5IdComponent")

Subsetting of an HDF5 dataset. The function writes an R data object to a subset of an HDF5 dataset. The given dimensions have to fit the dimensions of the HDF5 dataset. The HDF5 dataset has to be created beforehand, e.g. by h5createDataset().

[Package rhdf5 version 2.26.0 Index]