HDF5 Dataset Create Property List Interface

The functions, macros, and subroutines listed here are used to manipulate dataset creation property list objects in various ways, including to reset property values. With the use of property lists, HDF5 functions have been implemented and can be used in applications with many fewer parameters than would be required without property lists.

H5Pset_layout              ( h5plist, layout = h5default("H5D") )
H5Pget_layout              ( h5plist )
H5Pset_chunk               ( h5plist, dim )
H5Pget_chunk               ( h5plist )
H5Pset_deflate             ( h5plist, level )
H5Pset_fill_value          ( h5plist, value )
<!-- % H5Pget_fill_value                  ( plist_id, type_id, value ) -->
H5Pfill_value_defined      ( h5plist )
H5Pset_fill_time           ( h5plist, fill_time = h5default("H5D_FILL_TIME") )
H5Pget_fill_time           ( h5plist )
H5Pset_alloc_time          ( h5plist, alloc_time = h5default("H5D_ALLOC_TIME") )
H5Pget_alloc_time          ( h5plist )
<!-- % H5Pset_filter                      ( plist_id, filter_id, flags, cd_nelmts, cd_values[] ) -->
H5Pall_filters_avail       ( h5plist )
<!-- % H5Pget_nfilters                    ( plist ) -->
<!-- % H5Pget_filter1                     ( plist_id, idx, flags, cd_nelmts, cd_values, namelen, name[] ) -->
<!-- % H5Pget_filter2                     ( plist_id, idx, flags, cd_nelmts, cd_values[], namelen, name[], filter_config ) -->
<!-- % H5Pget_filter_by_id1               ( plist_id, filter_id, flags, cd_nelmts, cd_values[], namelen, name[] ) -->
<!-- % H5Pget_filter_by_id2               ( plist_id, filter_id, flags, cd_nelmts, cd_values[], namelen, name[], filter_config ) -->
<!-- % H5Pmodify_filter                   ( plist_id, filter_id, flags, cd_nelmts, cd_values[] ) -->
<!-- % H5Premove_filter                   ( plist_id, filter ) -->
<!-- % H5Pset_fletcher32                  ( plist_id ) -->
<!-- % H5Pset_nbit                        ( plist_id ) -->
<!-- % H5Pset_scaleoffset                 ( plist_id, scale_type, scale_factor ) -->
H5Pset_shuffle             ( h5plist )
H5Pset_szip                ( h5plist, options_mask, pixels_per_block )
<!-- % H5Pset_external          ( plist, *name, offset, size ) -->
<!-- % H5Pget_external_count    ( plist ) -->
<!-- % H5Pget_external                    ( plist, idx, name_size, *name, *offset, *size ) -->
H5Pset_blosc               ( h5plist, h5tid, method = 1L, level = 6L, shuffle = TRUE )
H5Pset_bzip2               ( h5plist, level = 2L )  
H5Pset_lzf                 ( h5plist, h5tid )

Arguments

h5plist
An object of class H5IdComponent representing a H5 property list identifier of class H5P_DATASET_CREATE. See H5Pcreate or H5Pcopy to create an object of this kind.

layout
A character name of a dataset layout type. See h5const("H5D") for possible property list types. Can also be an integer representing an HDF5 dataset layout type.

dim
The chunk size used to store the dataset. It is an integer vector of the same length as the dataset dims.

level
The compression level used. An integer value between 0 (no compression) and 9 (highest and slowest compression).

value
Standard value for filling the dataset. The storage.mode of value has to be convertable to the dataset type by HDF5.

fill_time
A character name of a H5D_FILL_TIME type. See h5const("H5D_FILL_TIME") for possible values. Can also be an integer representing an HDF5 H5D_FILL_TIME type.

alloc_time
A character name of a H5D_ALLOC_TIME type. See h5const("H5D_ALLOC_TIME") for possible property list types. Can also be an integer representing an HDF5 H5D_ALLOC_TIME type.

h5tid
HDF5 type ID.

method
When using the BLOSC filter, defines which compression tool should be used.

shuffle
Logical value determining whether byte shuffling should be applied before compression is applied. This only applies to the BLOSC filter.

options_mask
See https://portal.hdfgroup.org/display/HDF5/H5P_SET_SZIP for more details.

pixels_per_block
See https://portal.hdfgroup.org/display/HDF5/H5P_SET_SZIP for more details.

Details

Interface to the HDF5 C-library libhdf5. See https://portal.hdfgroup.org/display/HDF5/Dataset+Creation+Properties for further details. See H5P for documentation of more functions to manupulate property lists of other property list classes.

Value

The functions return the standard return value from their respective C-functions.

References

https://portal.hdfgroup.org/display/HDF5, https://portal.hdfgroup.org/display/HDF5/Szip+Compression+in+HDF+Products

Author

Bernd Fischer, Mike L. Smith

See also

Examples

pid <- H5Pcreate("H5P_DATASET_CREATE")
H5Pset_fill_time( pid, "H5D_FILL_TIME_ALLOC" )
H5Pset_chunk(pid, c(1000,1,1))
H5Pset_deflate(pid, 6)

H5Pget_layout( pid )
#> [1] "H5D_CHUNKED"
H5Pfill_value_defined( pid )
#> [1] TRUE
H5Pget_fill_time( pid )
#> [1] "H5D_FILL_TIME_ALLOC"
H5Pget_alloc_time( pid )
#> [1] "H5D_ALLOC_TIME_INCR"
H5Pget_chunk( pid)
#> [1]    1    1 1000

H5Pclose(pid)