HDF5 Datatype Interface

These functions create and manipulate the datatype which describes elements of a dataset.

H5Tcopy      (dtype_id = h5default(type = "H5T"))
H5Tset_cset  (dtype_id, cset = "ASCII")
H5Tget_cset  (dtype_id)
H5Tset_size  (dtype_id = h5default(type = "H5T"), size)
H5Tget_size  (dtype_id)
H5Tset_strpad(dtype_id, strpad = "NULLPAD")
H5Tget_strpad(dtype_id)

Arguments

dtype_id
A character name of a datatype. See h5const("H5T") for possible datatypes. Can also be an integer representing an HDF5 datatype.

cset
Encoding to use for string types. Valid options are 'ASCII' and 'UTF8'.

size
The total size in bytes.

strpad
Character string repsenting the type of padding to be used. Valid options are 'NULLTERM', 'NULLPAD', 'SPACEPAD'

Details

Interface to the HDF5 C-library libhdf5. See https://portal.hdfgroup.org/display/HDF5/Datatypes for further details.

Value

The functions return the standard return value from their respective C-functions.

References

https://portal.hdfgroup.org/display/HDF5

Author

Bernd Fischer

See also

Examples

# create character datatype with string length 10
tid <- H5Tcopy("H5T_C_S1")
H5Tset_size(tid, 10L)

# List all predefined types implemented in the R-interface
h5const("H5T")

#>  [1] "H5T_IEEE_F32BE"          "H5T_IEEE_F32LE"         
#>  [3] "H5T_IEEE_F64BE"          "H5T_IEEE_F64LE"         
#>  [5] "H5T_STD_I8BE"            "H5T_STD_I8LE"           
#>  [7] "H5T_STD_I16BE"           "H5T_STD_I16LE"          
#>  [9] "H5T_STD_I32BE"           "H5T_STD_I32LE"          
#> [11] "H5T_STD_I64BE"           "H5T_STD_I64LE"          
#> [13] "H5T_STD_U8BE"            "H5T_STD_U8LE"           
#> [15] "H5T_STD_U16BE"           "H5T_STD_U16LE"          
#> [17] "H5T_STD_U32BE"           "H5T_STD_U32LE"          
#> [19] "H5T_STD_U64BE"           "H5T_STD_U64LE"          
#> [21] "H5T_STD_B8BE"            "H5T_STD_B8LE"           
#> [23] "H5T_STD_B16BE"           "H5T_STD_B16LE"          
#> [25] "H5T_STD_B32BE"           "H5T_STD_B32LE"          
#> [27] "H5T_STD_B64BE"           "H5T_STD_B64LE"          
#> [29] "H5T_NATIVE_CHAR"         "H5T_NATIVE_SCHAR"       
#> [31] "H5T_NATIVE_UCHAR"        "H5T_NATIVE_SHORT"       
#> [33] "H5T_NATIVE_USHORT"       "H5T_NATIVE_INT"         
#> [35] "H5T_NATIVE_UINT"         "H5T_NATIVE_LONG"        
#> [37] "H5T_NATIVE_ULONG"        "H5T_NATIVE_LLONG"       
#> [39] "H5T_NATIVE_ULLONG"       "H5T_NATIVE_FLOAT"       
#> [41] "H5T_NATIVE_DOUBLE"       "H5T_NATIVE_LDOUBLE"     
#> [43] "H5T_NATIVE_B8"           "H5T_NATIVE_B16"         
#> [45] "H5T_NATIVE_B32"          "H5T_NATIVE_B64"         
#> [47] "H5T_NATIVE_OPAQUE"       "H5T_NATIVE_HADDR"       
#> [49] "H5T_NATIVE_HSIZE"        "H5T_NATIVE_HSSIZE"      
#> [51] "H5T_NATIVE_HERR"         "H5T_NATIVE_HBOOL"       
#> [53] "H5T_NATIVE_INT8"         "H5T_NATIVE_UINT8"       
#> [55] "H5T_NATIVE_INT_LEAST8"   "H5T_NATIVE_UINT_LEAST8" 
#> [57] "H5T_NATIVE_INT_FAST8"    "H5T_NATIVE_UINT_FAST8"  
#> [59] "H5T_NATIVE_INT16"        "H5T_NATIVE_UINT16"      
#> [61] "H5T_NATIVE_INT_LEAST16"  "H5T_NATIVE_UINT_LEAST16"
#> [63] "H5T_NATIVE_INT_FAST16"   "H5T_NATIVE_UINT_FAST16" 
#> [65] "H5T_NATIVE_INT32"        "H5T_NATIVE_UINT32"      
#> [67] "H5T_NATIVE_INT_LEAST32"  "H5T_NATIVE_UINT_LEAST32"
#> [69] "H5T_NATIVE_INT_FAST32"   "H5T_NATIVE_UINT_FAST32" 
#> [71] "H5T_NATIVE_INT64"        "H5T_NATIVE_UINT64"      
#> [73] "H5T_NATIVE_INT_LEAST64"  "H5T_NATIVE_UINT_LEAST64"
#> [75] "H5T_NATIVE_INT_FAST64"   "H5T_NATIVE_UINT_FAST64" 
#> [77] "H5T_NATIVE_DOUBLE"       "H5T_C_S1"               
#> [79] "H5T_FORTRAN_S1"         

# List all available type classes (not all of them are implemented)
h5const("H5T_CLASS")

#>  [1] "H5T_INTEGER"   "H5T_FLOAT"     "H5T_TIME"      "H5T_STRING"   
#>  [5] "H5T_BITFIELD"  "H5T_OPAQUE"    "H5T_COMPOUND"  "H5T_REFERENCE"
#>  [9] "H5T_ENUM"      "H5T_VLEN"      "H5T_ARRAY"