h5closeAll {rhdf5}	R Documentation

Close all open HDF5 handles

Description

Occasionally references to HDF5 files, groups, datasets etc can be created and not closed correctly. This function identifies all open handles and closes them. It replaces the functionality previously supplied by H5close.

Usage

h5closeAll()

Author(s)

Mike Smith

Examples

## create an empty file and then re-open it
h5createFile("ex_h5closeAll.h5")
H5Fopen("ex_h5closeAll.h5")

## list all open identifiers
h5listIdentifier()

## close all open identifiers and verify
h5closeAll()
h5listIdentifier()

[Package rhdf5 version 2.26.0 Index]