Close all open HDF5 handles

Occasionally references to HDF5 files, groups, datasets etc can be created and not closed correctly. This function identifies all open handles and closes them. It replaces the functionality previously supplied by H5close().

h5closeAll()

Value

Doesn't return anything. Called for the side-effect of closing any open HDF5 handles.

Author

Mike Smith

Examples

## create an empty file and then re-open it
h5createFile("ex_h5closeAll.h5")
#> file '/private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpzTdq9G/filebf77fb7d479/reference/ex_h5closeAll.h5' already exists.
H5Fopen("ex_h5closeAll.h5")

## list all open identifiers
h5listIdentifier()
#>       type name
#> 1 H5I_FILE    /

## close all open identifiers and verify
h5closeAll()
h5listIdentifier()
#> [1] type name
#> <0 rows> (or 0-length row.names)