h5ls {rhdf5}	R Documentation

List the content of an HDF5 file.

Description

Lists the content of an HDF5 file.

Usage

h5ls   (file,
        recursive = TRUE,
	all = FALSE, 
	datasetinfo = TRUE,
	index_type = h5default("H5_INDEX"),
	order = h5default("H5_ITER"), native = FALSE)
h5dump (file,
        recursive = TRUE,
	load = TRUE,
	all = FALSE,
        index_type = h5default("H5_INDEX"),
        order = h5default("H5_ITER"), ..., native = FALSE)

Arguments

file
The filename (character) of the file in which the dataset will be located. For advanced programmers it is possible to provide an object of class H5IdComponent representing a H5 location identifier (file or group). See H5Fcreate, H5Fopen, H5Gcreate, H5Gopen to create an object of this kind.

recursive
If TRUE, the content of the whole group hierarchy is listed. If FALSE, Only the content of the main group is shown. If a positive integer is provided this indicates the maximum level of the hierarchy that is shown.

all
If TRUE, a longer list of information on each entry is provided.

datasetinfo
If FALSE, datatype and dimensionality information is not provided. This can speed up the content listing for large files.

index_type
See h5const("H5_INDEX") for possible arguments.

order
See h5const("H5_ITER") for possible arguments.

load
If TRUE the datasets are read in, not only the header information. Note, that this can cause memory problems for very large files. In this case choose load=FALSE and load the datasets successively.

native
An object of class logical. If TRUE, array-like objects are treated as stored in HDF5 row-major rather than R column-major orientation. Using native = TRUE increases HDF5 file portability between programming languages. A file written with native = TRUE should also be read with native = TRUE

.

...
Arguments passed to h5read

Details

h5ls lists the content of an HDF5 file including group structure and datasets. It returns the content as a data.frame. You can use h5dump(file="myfile.h5", load=FALSE) to obtain the dataset information in a hierarchical list structure. Usually the datasets are loaded individually with h5read, but you have the possibility to load the complete content of an HDF5 file with h5dump

Value

h5ls returns a data.frame with the file content.

h5dump returns a hierarchical list structure representing the HDF5 group hierarchy. It either returns the datasets within the list structure (load=TRUE) or it returns a data.frame for each datset with the dataset header information load=FALSE.

Author(s)

Bernd Fischer

References

http://www.hdfgroup.org/HDF5

See Also

h5read, h5write, rhdf5

Examples

h5createFile("ex_ls_dump.h5")

# create groups
h5createGroup("ex_ls_dump.h5","foo")
h5createGroup("ex_ls_dump.h5","foo/foobaa")

# write a matrix
B = array(seq(0.1,2.0,by=0.1),dim=c(5,2,2))
attr(B, "scale") <- "liter"
h5write(B, "ex_ls_dump.h5","foo/B")

# list content of hdf5 file
h5ls("ex_ls_dump.h5",all=TRUE)
h5dump("ex_ls_dump.h5")

[Package rhdf5 version 2.26.0 Index]