List the content of an HDF5 file.

List the content of an HDF5 file.

h5ls(
  file,
  recursive = TRUE,
  all = FALSE,
  datasetinfo = TRUE,
  index_type = h5default("H5_INDEX"),
  order = h5default("H5_ITER"),
  s3 = FALSE,
  s3credentials = NULL,
  native = FALSE
)

Arguments

file
The filename (character) of the file in which the dataset will be located. You can also provide an object of class H5IdComponent representing a H5 location identifier (file or group). See H5Fcreate(), H5Fopen(), H5Gcreate(), H5Gopen() to create an object of this kind.

recursive
If TRUE, the content of the whole group hierarchy is listed. If FALSE, Only the content of the main group is shown. If a positive integer is provided this indicates the maximum level of the hierarchy that is shown.

all
If TRUE, a longer list of information on each entry is provided.

datasetinfo
If FALSE, datatype and dimensionality information is not provided. This can speed up the content listing for large files.

index_type
See h5const("H5_INDEX") for possible arguments.

order
See h5const("H5_ITER") for possible arguments.

s3
Logical value indicating whether the file argument should be treated as a URL to an Amazon S3 bucket, rather than a local file path.

s3credentials
A list of length three, providing the credentials for accessing files in a private Amazon S3 bucket.

native
An object of class logical. If TRUE, array-like objects are treated as stored in HDF5 row-major rather than R column-major orientation. Using native = TRUE increases HDF5 file portability between programming languages. A file written with native = TRUE should also be read with native = TRUE

Value

h5ls returns a data.frame with the file content.

References

https://portal.hdfgroup.org/display/HDF5

See also

Author

Bernd Fischer, Mike L. Smith

Examples

h5createFile("ex_ls_dump.h5")
#> file '/private/var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/RtmpTZxdvV/fileb844e4d77f5/reference/ex_ls_dump.h5' already exists.

# create groups
h5createGroup("ex_ls_dump.h5","foo")
#> Can not create group. Object with name 'foo' already exists.
#> [1] FALSE
h5createGroup("ex_ls_dump.h5","foo/foobaa")
#> Can not create group. Object with name 'foo/foobaa' already exists.
#> [1] FALSE

# write a matrix
B = array(seq(0.1,2.0,by=0.1),dim=c(5,2,2))
attr(B, "scale") <- "liter"
h5write(B, "ex_ls_dump.h5","foo/B")

# list content of hdf5 file
h5ls("ex_ls_dump.h5",all=TRUE)
#>   group   name         ltype cset       otype num_attrs dclass          dtype
#> 0     /    foo H5L_TYPE_HARD    0   H5I_GROUP         0                      
#> 1  /foo      B H5L_TYPE_HARD    0 H5I_DATASET         1  FLOAT H5T_IEEE_F64LE
#> 2  /foo foobaa H5L_TYPE_HARD    0   H5I_GROUP         0                      
#>    stype rank       dim    maxdim
#> 0           0                    
#> 1 SIMPLE    3 5 x 2 x 2 5 x 2 x 2
#> 2           0                    
h5dump("ex_ls_dump.h5")
#> $foo
#> $foo$B
#> , , 1
#> 
#>      [,1] [,2]
#> [1,]  0.1  0.6
#> [2,]  0.2  0.7
#> [3,]  0.3  0.8
#> [4,]  0.4  0.9
#> [5,]  0.5  1.0
#> 
#> , , 2
#> 
#>      [,1] [,2]
#> [1,]  1.1  1.6
#> [2,]  1.2  1.7
#> [3,]  1.3  1.8
#> [4,]  1.4  1.9
#> [5,]  1.5  2.0
#> 
#> 
#> $foo$foobaa
#> NULL
#> 
#> 

# list content of an hdf5 file in a public S3 bucket
# \donttest{
h5ls(file = "https://rhdf5-public.s3.eu-central-1.amazonaws.com/h5ex_t_array.h5", s3 = TRUE)
#> Error in H5Pset_fapl_ros3(fapl, s3credentials) Rhdf5lib was not compiled with support for the S3 VFD
# }